Fixtures and pytest config cleanup
